requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.0

# LLM and AI
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.5

# Vector database
chromadb==0.4.18

# Embeddings and ML
sentence-transformers==2.2.2

# Data handling
pydantic==2.5.0
python-multipart==0.0.6

# File handling
python-dotenv==1.0.0
zipfile36==0.1.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-json-logger==2.0.7
typing-extensions==4.8.0